Add Grid Scaling Feature
After debugging for an hour, exploring documentation and flex properties, the grid can have MxN cells without exceeding the max width.

I learned the power of flex grow and its impact on empty containers. It was troublesome trying to find a percentage version of padding. I knew that this was possible, even if I did not know how. In the end, I figured it out.