Implement Extra Credit 2
Color slowly turns to black each time the mouse is over the element. The code is currently messy, but it worked. Basic RegEx was enough to extract the values of RGB.

I also learned that the match function returns the entire matched string and the output of each grouping function.

I will not be updating the design of the website  anymore though. I am eager to learn next rather than perfecting this. This activity forced me to explore documentation and ways to solve feature problems.